Fullstack/pdf_search_app/backend/requirements.txt
Flask
python-dotenv
opensearch-py
minio
PyPDF2
Flask-Cors
